Category

+++ src/main/java/com/revature/GroupDP2/model/Category.java
@@ -1,15 +1,16 @@
 package com.revature.GroupDP2.model;
 
+import org.json.JSONPropertyName;
+
 import javax.persistence.*;
 
 @Entity
 @Table(name = "categories", schema = "groupd")
 public class Category {
-
     @Id
     @Column(name = "category_id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long categoryId;
+    private Integer categoryId;
 
     @Column(name = "category_name")
     private String categoryName;
@@ -21,11 +22,11 @@ public class Category {
         this.categoryName = categoryName;
     }
 
-    public Long getCategoryId() {
+    public Integer getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(Long categoryId) {
+    public void setCategoryId(Integer categoryId) {
         this.categoryId = categoryId;
     }
